+++ NEWS.md
@@ -2,8 +2,30 @@
 
 ## New Features:
 
+* New age helpers `add_years()` and `age_on_date()` has been added to help compute features for individual data (#125).
+
+* Added the fields `$available_observables` and `$available_features` for easier overview (#139).
+
 * The data availability period for each `diseasystore` is now exposed via the `$min_start_date` and `$max_end_date`
 fields (#138).
+
+## Minor Improvements and Fixes:
+
+* Two bugs were fixed in `$determine_new_ranges()` where existing tables were not detected:
+  * On back ends that use "catalog" to structure table (DuckDB and SQL Server) (fix also requires SCDB > v0.4) (#158).
+  * Due to `slice_ts` not being correctly matched with existing data on some backends (#172).
+
+* Long stratification expression are now properly parsed in `$key_join_features()` (#161).
+
+## Testing:
+
+* `test_diseasystore()` now also checks that the `FeatureHandler`s return data directly: (#154)
+  * Checks that data is only within the study period.
+  * Checks that `valid_from` and `valid_until` has class `Date`.
+
+## Documentation:
+
+* Added benchmarking vignette `vignette("benchmarks")` (#144).
 
 
 ## Minor Improvements and Fixes: